a hole bunch of updates, here is a summary: major changes: - moved sysdata from develop/sysdata/ to sysdata/ - added many required imports and exports to the NAMESPACE and restructured the file for better readability - added missing data "foofit" (a twinSIR fit) to version control - added sts-version of "ha" data from Berlin to data/ for convenient use in the examples - new function ks.plot.unif (yet undocumented), which is used by residuals.twinSIR and residuals.twinstim (which are now registered S3 methods, but not yet completely documented) - added quasi-exact method for cubature of the bivariate normal density over polygonal domains; also added a wrapper function (polyCub) for the different cubature methods; furthermore, the cubature methods gained a plot-argument for requesting an illustrative plot of the cubature - removed Michaels crude bug fix in polyCub.midpoint as this might destroy convergence or accuracy; let's see if there is still need for some error catching - separated spatial_stuff.R and spatial_coerce-methods.R from twinstim_helper.R - added and improved documentation (still some man-pages and examples missing)
minor changes:
- moved package xtable from suggests to imports since we define an S3 method for xtable and because it is quiet nice to have the toLatex functionality per default (not only suggested)
- added variable for R executable in the Makefile which enables the use of, e.g., make check R=R-devel (if R-devel links to the binary of the development version of R somewhere on the computer)
- replaced exit() calls in twins.cc by Rf_error()
- replaced some more "long" types by "int" in C code for compatibility with 64bit architectures (this time in stcd-assuncaocorrea.cc)
- fixed the R CMD check warnings about partially matched argumentsin function calls
- made as.epidata S3 generic
- registered "owin" as old-style class
- updated my affiliation in www/index.php
- new internal auxiliary function dotprod



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/surveillance@283 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/src/stcd-assuncaocorrea.cc
@@ -283,7 +283,7 @@ int CalculaLambda(SVEventLst &ev, const double RaioC, const double epslon, std::
 
 extern "C" {
 
-  void SRspacetime(double *x, double *y, double *t, long *n, double *radius,
+  void SRspacetime(double *x, double *y, double *t, int *n, double *radius,
 		   double *epsilon, double *areaA, double *areaAcapBk,
 		   int *cusum, double *threshold, 
 		   double *Rarray, int *idxFirstAlarm, int *idxClusterCenter) {
@@ -292,7 +292,7 @@ extern "C" {
   SVEvent e;
   SVEventLst eList;
   unsigned int i;
-  long j;
+  int j;
   //Fill coordinates of event list
   for(j=0;j<*n;j++){
     e.x = x[j];
